исправляет функцию
